Add: Support for CUDA 7.0 Fix: 64-bit pointer for BLAS, SPARSE and RAND Chg: Remove x86 support for BLAS, SPARSE and RAND Chg: Make GPU architecture 2.0 the default Chg: Remove x86 versions of CudafyViewer.exe and cudafycl.exe
git-svn-id: https://cudafy.svn.codeplex.com/svn@29295 2b2adeab-ffde-484a-ad65-8eada40eade2

+++ Cudafy/3p/cuda.net3.0.0_win/src/CUDA.NET_3.0_Source/GASS.CUDA.BLAS.Types/cublasHandle.cs
@@ -6,7 +6,7 @@
     [StructLayout(LayoutKind.Sequential)]
     public struct cublasHandle
     {
-        public uint handle;
+        public ulong handle;
     }
 }
